+++ app/ui/navigation/index.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+import Navigation from "./navigation";
+
+export default Navigation;
